monglue: Introduce monglue-register
Monglue (for monitoring glue), is a simple web service (using sinatra),
used to register (in a redis db) all VMid to monitor for billing. When a
VM boot, the monitoring agent notify the monitoring service in order to
start the monitoring process, then the monitoring server send the VMid
to this service in order to request metrics (with this ID).

In monglue, we use a simple redis list with all VM id for a later use.
We use a redis db in order to be asynchrony with read/write requests (and
not block with a read/write in a file).