monglue: Introduce monglue-control (for ebill population)
monglue-control is a simple service to make the interface with the
monitoring service and the billing service. The redis database populated
by the monitoring service and monglue-register is readed by
monglue-control in order to request the corresponding metrics to the
monitoring service (GET), with the data requested we re-make an http
request (POST) to the billing service for historical and latter
processing.